# Conflicts:
#	src/test/scala/io/qbeast/core/transform/HistogramTransformationTest.scala
#	src/test/scala/io/qbeast/spark/index/TransformerIndexingTest.scala

+++ src/test/scala/io/qbeast/TestClasses.scala
@@ -16,11 +16,11 @@
 package io.qbeast
 
 object TestClasses {
-  case class T1(a: Int, b: String, c: Double)
+  case class T1(a: Long, b: String, c: Double)
 
-  case class T2(a: Int, c: Double)
+  case class T2(a: Long, c: Double)
 
-  case class T3(a: Int, b: Double, c: String, d: Float)
+  case class T3(a: Long, b: Double, c: String, d: Float)
 
   case class TestStrings(a: String, b: String, c: String)
   case class TestDouble(a: Double, b: Double, c: Double)
@@ -28,7 +28,7 @@ object TestClasses {
   case class TestBigDecimal(a: BigDecimal, b: BigDecimal, c: BigDecimal)
   case class TestInt(a: Int, b: Int, c: Int)
   case class TestLong(a: Long, b: Long, c: Long)
-  case class TestNull(a: Option[String], b: Option[Double], c: Option[Int])
+  case class TestNull(a: Option[String], b: Option[Double], c: Option[Long])
 
   case class IndexData(id: Long, cube: Array[Byte], weight: Double, state: String)
